backend/requirements.txt
# EcosystIA Backend Requirements
# Core Django
Django==5.0.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
django-allauth==0.57.0
cryptography==42.0.5
bcrypt==4.1.3

# Database
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0

# API & Serialization
django-filter==23.5
drf-spectacular==0.27.0
django-extensions==3.2.3

# AI & ML
google-generativeai==0.8.3
openai==1.12.0
celery==5.3.4

# File Handling & Storage
Pillow==10.2.0
django-storages==1.14.2
boto3==1.34.34

# Monitoring & Logging
sentry-sdk==1.39.2
django-health-check==3.17.0

# Development
pytest==8.0.0
pytest-django==4.7.0
black==24.2.0
flake8==7.0.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
